Add Comonad instance
